New files - small errors
NOTE: in the schematic, the UART interface to the ATMEGA328 UART pins is wrong.

The molex Tx should voltage divide to the 3.3V Rx of the ATMEGA, and a pull-up resistor + LED on the 3.3V segment

The ATMEGA Tx should pass through a 1K current-limiting resistor to the Molex Rx (no voltage division needed). A pull-up resistor and LED should be on the segment of wire between the ATMEGA and 1K resistor.

The parts list does not accurately reflect the current state and needs to be updated.