Small change to screws
Modified them to 2.2mm in diameter. Seems the fab house changed them from 3mm to 2.2mm in the order, which actually worked out better for fitting screws in.

The 'top-left' screw (between pin header and ICSP header) was moved 0.3mm to the right to allow more of the ground plane to connect. Will be reflected in the 3D model of the casing.